+++ Thoughts/Extensions/Color.swift
@@ -1,0 +1,360 @@
+// Copyright (c) 2024 Jason Morley
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+import SwiftUI
+
+import HashRainbow
+
+extension Color {
+
+    struct HTML {
+
+        public static let aliceBlue = Color(0xF0F8FF)
+        public static let antiqueWhite = Color(0xFAEBD7)
+        public static let aqua = Color(0x00FFFF)
+        public static let aquamarine = Color(0x7FFFD4)
+        public static let azure = Color(0xF0FFFF)
+        public static let beige = Color(0xF5F5DC)
+        public static let bisque = Color(0xFFE4C4)
+        public static let black = Color(0x000000)
+        public static let blanchedAlmond = Color(0xFFEBCD)
+        public static let blue = Color(0x0000FF)
+        public static let blueViolet = Color(0x8A2BE2)
+        public static let brown = Color(0xA52A2A)
+        public static let burlyWood = Color(0xDEB887)
+        public static let cadetBlue = Color(0x5F9EA0)
+        public static let chartreuse = Color(0x7FFF00)
+        public static let chocolate = Color(0xD2691E)
+        public static let coral = Color(0xFF7F50)
+        public static let cornflowerBlue = Color(0x6495ED)
+        public static let cornsilk = Color(0xFFF8DC)
+        public static let crimson = Color(0xDC143C)
+        public static let cyan = Color(0x00FFFF)
+        public static let darkBlue = Color(0x00008B)
+        public static let darkCyan = Color(0x008B8B)
+        public static let darkGoldenRod = Color(0xB8860B)
+        public static let darkGray = Color(0xA9A9A9)
+        public static let darkGrey = Color(0xA9A9A9)
+        public static let darkGreen = Color(0x006400)
+        public static let darkKhaki = Color(0xBDB76B)
+        public static let darkMagenta = Color(0x8B008B)
+        public static let darkOliveGreen = Color(0x556B2F)
+        public static let darkOrange = Color(0xFF8C00)
+        public static let darkOrchid = Color(0x9932CC)
+        public static let darkRed = Color(0x8B0000)
+        public static let darkSalmon = Color(0xE9967A)
+        public static let darkSeaGreen = Color(0x8FBC8F)
+        public static let darkSlateBlue = Color(0x483D8B)
+        public static let darkSlateGray = Color(0x2F4F4F)
+        public static let darkSlateGrey = Color(0x2F4F4F)
+        public static let darkTurquoise = Color(0x00CED1)
+        public static let darkViolet = Color(0x9400D3)
+        public static let deepPink = Color(0xFF1493)
+        public static let deepSkyBlue = Color(0x00BFFF)
+        public static let dimGray = Color(0x696969)
+        public static let dimGrey = Color(0x696969)
+        public static let dodgerBlue = Color(0x1E90FF)
+        public static let fireBrick = Color(0xB22222)
+        public static let floralWhite = Color(0xFFFAF0)
+        public static let forestGreen = Color(0x228B22)
+        public static let fuchsia = Color(0xFF00FF)
+        public static let gainsboro = Color(0xDCDCDC)
+        public static let ghostWhite = Color(0xF8F8FF)
+        public static let gold = Color(0xFFD700)
+        public static let goldenRod = Color(0xDAA520)
+        public static let gray = Color(0x808080)
+        public static let grey = Color(0x808080)
+        public static let green = Color(0x008000)
+        public static let greenYellow = Color(0xADFF2F)
+        public static let honeyDew = Color(0xF0FFF0)
+        public static let hotPink = Color(0xFF69B4)
+        public static let indianRed  = Color(0xCD5C5C)
+        public static let indigo   = Color(0x4B0082)
+        public static let ivory = Color(0xFFFFF0)
+        public static let khaki = Color(0xF0E68C)
+        public static let lavender = Color(0xE6E6FA)
+        public static let lavenderBlush = Color(0xFFF0F5)
+        public static let lawnGreen = Color(0x7CFC00)
+        public static let lemonChiffon = Color(0xFFFACD)
+        public static let lightBlue = Color(0xADD8E6)
+        public static let lightCoral = Color(0xF08080)
+        public static let lightCyan = Color(0xE0FFFF)
+        public static let lightGoldenRodYellow = Color(0xFAFAD2)
+        public static let lightGray = Color(0xD3D3D3)
+        public static let lightGrey = Color(0xD3D3D3)
+        public static let lightGreen = Color(0x90EE90)
+        public static let lightPink = Color(0xFFB6C1)
+        public static let lightSalmon = Color(0xFFA07A)
+        public static let lightSeaGreen = Color(0x20B2AA)
+        public static let lightSkyBlue = Color(0x87CEFA)
+        public static let lightSlateGray = Color(0x778899)
+        public static let lightSlateGrey = Color(0x778899)
+        public static let lightSteelBlue = Color(0xB0C4DE)
+        public static let lightYellow = Color(0xFFFFE0)
+        public static let lime = Color(0x00FF00)
+        public static let limeGreen = Color(0x32CD32)
+        public static let linen = Color(0xFAF0E6)
+        public static let magenta = Color(0xFF00FF)
+        public static let maroon = Color(0x800000)
+        public static let mediumAquaMarine = Color(0x66CDAA)
+        public static let mediumBlue = Color(0x0000CD)
+        public static let mediumOrchid = Color(0xBA55D3)
+        public static let mediumPurple = Color(0x9370DB)
+        public static let mediumSeaGreen = Color(0x3CB371)
+        public static let mediumSlateBlue = Color(0x7B68EE)
+        public static let mediumSpringGreen = Color(0x00FA9A)
+        public static let mediumTurquoise = Color(0x48D1CC)
+        public static let mediumVioletRed = Color(0xC71585)
+        public static let midnightBlue = Color(0x191970)
+        public static let mintCream = Color(0xF5FFFA)
+        public static let mistyRose = Color(0xFFE4E1)
+        public static let moccasin = Color(0xFFE4B5)
+        public static let navajoWhite = Color(0xFFDEAD)
+        public static let navy = Color(0x000080)
+        public static let oldLace = Color(0xFDF5E6)
+        public static let olive = Color(0x808000)
+        public static let oliveDrab = Color(0x6B8E23)
+        public static let orange = Color(0xFFA500)
+        public static let orangeRed = Color(0xFF4500)
+        public static let orchid = Color(0xDA70D6)
+        public static let paleGoldenRod = Color(0xEEE8AA)
+        public static let paleGreen = Color(0x98FB98)
+        public static let paleTurquoise = Color(0xAFEEEE)
+        public static let paleVioletRed = Color(0xDB7093)
+        public static let papayaWhip = Color(0xFFEFD5)
+        public static let peachPuff = Color(0xFFDAB9)
+        public static let peru = Color(0xCD853F)
+        public static let pink = Color(0xFFC0CB)
+        public static let plum = Color(0xDDA0DD)
+        public static let powderBlue = Color(0xB0E0E6)
+        public static let purple = Color(0x800080)
+        public static let rebeccaPurple = Color(0x663399)
+        public static let red = Color(0xFF0000)
+        public static let rosyBrown = Color(0xBC8F8F)
+        public static let royalBlue = Color(0x4169E1)
+        public static let saddleBrown = Color(0x8B4513)
+        public static let salmon = Color(0xFA8072)
+        public static let sandyBrown = Color(0xF4A460)
+        public static let seaGreen = Color(0x2E8B57)
+        public static let seaShell = Color(0xFFF5EE)
+        public static let sienna = Color(0xA0522D)
+        public static let silver = Color(0xC0C0C0)
+        public static let skyBlue = Color(0x87CEEB)
+        public static let slateBlue = Color(0x6A5ACD)
+        public static let slateGray = Color(0x708090)
+        public static let slateGrey = Color(0x708090)
+        public static let snow = Color(0xFFFAFA)
+        public static let springGreen = Color(0x00FF7F)
+        public static let steelBlue = Color(0x4682B4)
+        public static let tan = Color(0xD2B48C)
+        public static let teal = Color(0x008080)
+        public static let thistle = Color(0xD8BFD8)
+        public static let tomato = Color(0xFF6347)
+        public static let turquoise = Color(0x40E0D0)
+        public static let violet = Color(0xEE82EE)
+        public static let wheat = Color(0xF5DEB3)
+        public static let white = Color(0xFFFFFF)
+        public static let whiteSmoke = Color(0xF5F5F5)
+        public static let yellow = Color(0xFFFF00)
+        public static let yellowGreen = Color(0x9ACD32)
+
+    }
+
+    init(_ value: Int32) {
+        let red = Double(((0xff << 16) & value) >> 16)
+        let green = Double(((0xff << 8) & value) >> 8)
+        let blue = Double(0xff & value)
+        self.init(red: red / 255, green: green / 255, blue: blue / 255)
+    }
+
+    var gray: CGFloat? {
+        return cgColor?.converted(to: CGColorSpaceCreateDeviceGray(),
+                                  intent: CGColorRenderingIntent.defaultIntent,
+                                  options: nil)?.components?.first
+    }
+
+}
+
+extension Array where Element == Color {
+
+    static let html: [Color] = [
+        .HTML.aliceBlue,
+        .HTML.antiqueWhite,
+        .HTML.aqua,
+        .HTML.aquamarine,
+        .HTML.azure,
+        .HTML.beige,
+        .HTML.bisque,
+        .HTML.black,
+        .HTML.blanchedAlmond,
+        .HTML.blue,
+        .HTML.blueViolet,
+        .HTML.brown,
+        .HTML.burlyWood,
+        .HTML.cadetBlue,
+        .HTML.chartreuse,
+        .HTML.chocolate,
+        .HTML.coral,
+        .HTML.cornflowerBlue,
+        .HTML.cornsilk,
+        .HTML.crimson,
+        .HTML.cyan,
+        .HTML.darkBlue,
+        .HTML.darkCyan,
+        .HTML.darkGoldenRod,
+        .HTML.darkGray,
+        .HTML.darkGrey,
+        .HTML.darkGreen,
+        .HTML.darkKhaki,
+        .HTML.darkMagenta,
+        .HTML.darkOliveGreen,
+        .HTML.darkOrange,
+        .HTML.darkOrchid,
+        .HTML.darkRed,
+        .HTML.darkSalmon,
+        .HTML.darkSeaGreen,
+        .HTML.darkSlateBlue,
+        .HTML.darkSlateGray,
+        .HTML.darkSlateGrey,
+        .HTML.darkTurquoise,
+        .HTML.darkViolet,
+        .HTML.deepPink,
+        .HTML.deepSkyBlue,
+        .HTML.dimGray,
+        .HTML.dimGrey,
+        .HTML.dodgerBlue,
+        .HTML.fireBrick,
+        .HTML.floralWhite,
+        .HTML.forestGreen,
+        .HTML.fuchsia,
+        .HTML.gainsboro,
+        .HTML.ghostWhite,
+        .HTML.gold,
+        .HTML.goldenRod,
+        .HTML.gray,
+        .HTML.grey,
+        .HTML.green,
+        .HTML.greenYellow,
+        .HTML.honeyDew,
+        .HTML.hotPink,
+        .HTML.indianRed,
+        .HTML.indigo,
+        .HTML.ivory,
+        .HTML.khaki,
+        .HTML.lavender,
+        .HTML.lavenderBlush,
+        .HTML.lawnGreen,
+        .HTML.lemonChiffon,
+        .HTML.lightBlue,
+        .HTML.lightCoral,
+        .HTML.lightCyan,
+        .HTML.lightGoldenRodYellow,
+        .HTML.lightGray,
+        .HTML.lightGrey,
+        .HTML.lightGreen,
+        .HTML.lightPink,
+        .HTML.lightSalmon,
+        .HTML.lightSeaGreen,
+        .HTML.lightSkyBlue,
+        .HTML.lightSlateGray,
+        .HTML.lightSlateGrey,
+        .HTML.lightSteelBlue,
+        .HTML.lightYellow,
+        .HTML.lime,
+        .HTML.limeGreen,
+        .HTML.linen,
+        .HTML.magenta,
+        .HTML.maroon,
+        .HTML.mediumAquaMarine,
+        .HTML.mediumBlue,
+        .HTML.mediumOrchid,
+        .HTML.mediumPurple,
+        .HTML.mediumSeaGreen,
+        .HTML.mediumSlateBlue,
+        .HTML.mediumSpringGreen,
+        .HTML.mediumTurquoise,
+        .HTML.mediumVioletRed,
+        .HTML.midnightBlue,
+        .HTML.mintCream,
+        .HTML.mistyRose,
+        .HTML.moccasin,
+        .HTML.navajoWhite,
+        .HTML.navy,
+        .HTML.oldLace,
+        .HTML.olive,
+        .HTML.oliveDrab,
+        .HTML.orange,
+        .HTML.orangeRed,
+        .HTML.orchid,
+        .HTML.paleGoldenRod,
+        .HTML.paleGreen,
+        .HTML.paleTurquoise,
+        .HTML.paleVioletRed,
+        .HTML.papayaWhip,
+        .HTML.peachPuff,
+        .HTML.peru,
+        .HTML.pink,
+        .HTML.plum,
+        .HTML.powderBlue,
+        .HTML.purple,
+        .HTML.rebeccaPurple,
+        .HTML.red,
+        .HTML.rosyBrown,
+        .HTML.royalBlue,
+        .HTML.saddleBrown,
+        .HTML.salmon,
+        .HTML.sandyBrown,
+        .HTML.seaGreen,
+        .HTML.seaShell,
+        .HTML.sienna,
+        .HTML.silver,
+        .HTML.skyBlue,
+        .HTML.slateBlue,
+        .HTML.slateGray,
+        .HTML.slateGrey,
+        .HTML.snow,
+        .HTML.springGreen,
+        .HTML.steelBlue,
+        .HTML.tan,
+        .HTML.teal,
+        .HTML.thistle,
+        .HTML.tomato,
+        .HTML.turquoise,
+        .HTML.violet,
+        .HTML.wheat,
+        .HTML.white,
+        .HTML.whiteSmoke,
+        .HTML.yellow,
+        .HTML.yellowGreen,
+    ]
+
+    static let tags: [Color] = {
+        return html.filter { color in
+            guard let gray = color.gray,
+                  gray < 0.9,
+                  gray > 0.1
+            else {
+                return false
+            }
+            return true
+        }
+    }()
+
+}
